requirements.txt
# Core dependencies (Python 3.8 compatible)
numpy>=1.19.0,<1.25.0  # Python 3.8 compatible
opencv-python>=4.5.0
scikit-image>=0.18.0
scipy>=1.7.0,<1.11.0  # Python 3.8 max is 1.10.1
pandas>=1.3.0,<2.0.0  # Python 3.8 compatible

# Machine Learning
ultralytics>=8.0.0  # For YOLOv8-seg
# torch>=2.0.0  # PyTorch (installed separately, often via conda)
# torchvision>=0.15.0

# Optional: For Detectron2/Mask R-CNN
# detectron2  # Install separately: pip install detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/torch2.0/index.html

# Cost estimation ML models
xgboost>=1.6.0  # Python 3.8 compatible
lightgbm>=3.3.0  # Python 3.8 compatible
scikit-learn>=1.0.0

# Data handling
pyyaml>=6.0
# pycocotools>=2.0.6  # For COCO dataset handling

# Optional: For depth estimation (if using)
# transformers>=4.30.0  # For MiDaS depth models

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

